Move maps tests to own file
